CORE-5271 - refactor DatabaseChunkPersistence.kt, split into two interfaces for chunk and cpi persistence.
